Moved config to a folder
