+++ client/src/components/landingPage.js
@@ -3,7 +3,7 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import "./CSS/landingpage.css";
 import axios from "axios";
-const urls = require("../config.json");
+const urls = require("../config/config.json");
 
 class LandingPage extends Component {
   componentDidMount() {
