+++ client/src/components/register.js
@@ -10,7 +10,7 @@ import {
 import "./CSS/login.css";
 import axios from "axios";
 
-const urls = require("../config.json");
+const urls = require("../config/config.json");
 
 class Register extends Component {
   state = {
