+++ client/src/components/skillsCreate.js
@@ -4,7 +4,7 @@ import Sidebar from "./subComponents/sidebar";
 import axios from "axios";
 import Navbar from "./subComponents/navbar";
 
-const urls = require("../config.json");
+const urls = require("../config/config.json");
 
 class SkillsCreate extends Component {
   constructor(props) {
